Fixes to HTTP parser lib, DTSCMeta keysizes, MP4 headers and 4+GiB MP4 file output by Oswald de Bruin

+++ lib/http_parser.cpp
@@ -309,10 +309,10 @@ void HTTP::Parser::SetHeader(std::string i, std::string v) {
 }
 
 /// Sets header i to integer value v.
-void HTTP::Parser::SetHeader(std::string i, int v) {
+void HTTP::Parser::SetHeader(std::string i, long long v) {
   Trim(i);
   char val[23]; //ints are never bigger than 22 chars as decimal
-  sprintf(val, "%i", v);
+  sprintf(val, "%lld", v);
   headers[i] = val;
 }
